+++ R/stat-dens2d-labels.r
@@ -8,12 +8,33 @@
 #'   to \code{rownames(data)}, with a message.
 #'
 #' @details \code{stat_dens2d_labels()} is designed to work together with
-#'   statistics from package 'ggrepel'. To avoid text labels being plotted over
+#'   geometries from packages 'ggrepel'. To avoid text labels being plotted over
 #'   unlabelled points the corresponding rows in data need to be retained but
-#'   labels replaced with the empty character string, \code{""}. This makes
-#'   \code{\link{stat_dens2d_filter}} unsuitable for the task. Non-the-less
-#'   \code{stat_dens2d_labels()} could be useful in some other cases, as the
+#'   labels replaced with the empty character string, \code{""}. Function
+#'   \code{\link{stat_dens2d_filter}} cannot be used with the repulsive geoms
+#'   from 'ggrepel' because it drops the observations. Non-the-less
+#'   \code{stat_dens2d_labels()} can be useful in some other cases, as the
 #'   substitution character string can be set by the user.
+#'
+#'   The local density of observations in 2D (\emph{x} and \emph{y}) is computed
+#'   with function \code{\link[MASS]{kde2d}} and used to select observations,
+#'   passing to the geom all the rows in its \code{data} input but with with the
+#'   text of labels replaced in those "not kept". The default is to select
+#'   observations in sparse regions of the plot, but the selection can be
+#'   inverted so that only observations in the densest regions are returned.
+#'   Specific observations can be protected from having the label replaced by
+#'   passing a suitable argument to \code{keep.these}. Logical and integer
+#'   vectors function as indexes to rows in \code{data}, while a character
+#'   vector is compared to values in the variable mapped to the \code{label}
+#'   aesthetic. A function passed as argument to keep.these will receive as
+#'   argument the values in the variable mapped to \code{label} and should
+#'   return a character, logical or numeric vector as described above.
+#'
+#'   How many labels are retained intact in addition to those in
+#'   \code{keep.these} is controlled with arguments passed to \code{keep.number}
+#'   and \code{keep.fraction}. \code{keep.number} sets the maximum number of
+#'   observations selected, whenever \code{keep.fraction} results in fewer
+#'   observations selected, it is obeyed.
 #'
 #' @param mapping The aesthetic mapping, usually constructed with
 #'   \code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs
@@ -29,6 +50,10 @@
 #' @param keep.sparse logical If \code{TRUE}, the default, observations from the
 #'   more sparse regions are retained, if \code{FALSE} those from the densest
 #'   regions.
+#' @param keep.these character vector, integer vector, logical vector or
+#'   function that takes the variable mapped to the \code{label} aesthetic as
+#'   first argument and returns a character vector or a logical vector. These
+#'   rows from \code{data} are selected irrespective of the local density.
 #' @param invert.selection logical If \code{TRUE}, the complement of the
 #'   selected rows are returned.
 #' @param h vector of bandwidths for x and y directions. Defaults to normal
@@ -55,7 +80,10 @@
 #' @return A plot layer instance. Using as output \code{data} the input
 #'   \code{data} after value substitution based on a 2D the filtering criterion.
 #'
-#' @seealso \code{\link[MASS]{kde2d}} used internally.
+#' @seealso \code{\link{stat_dens2d_filter}} and \code{\link[MASS]{kde2d}} used
+#'   internally. Parameters \code{n}, \code{h} in this statistic correspond to
+#'   the parameters with the same name in this imported function. Limits are set
+#'   to the limits of the plot scales.
 #'
 #' @family statistics returning a subset of data
 #'
@@ -81,6 +109,14 @@
 #' ggplot(data = d, aes(x, y, label = lab)) +
 #'   geom_point() +
 #'   stat_dens2d_labels()
+#'
+#' ggplot(data = d, aes(x, y, label = lab)) +
+#'   geom_point() +
+#'   stat_dens2d_labels(keep.these = "zoujdg")
+#'
+#' ggplot(data = d, aes(x, y, label = lab)) +
+#'   geom_point() +
+#'   stat_dens2d_labels(keep.these = function(x) {grepl("^z", x)})
 #'
 #' ggplot(data = d, aes(x, y, label = lab)) +
 #'   geom_point() +
@@ -122,6 +158,10 @@
 #'   ggplot(data = d, aes(x, y, label = lab)) +
 #'     geom_point() +
 #'     stat_dens2d_labels(geom = "debug")
+#'
+#'   ggplot(data = d, aes(x, y)) +
+#'     geom_point() +
+#'     stat_dens2d_labels(geom = "debug")
 #' }
 #'
 stat_dens2d_labels <-
@@ -133,6 +173,7 @@ stat_dens2d_labels <-
            keep.fraction = 0.10,
            keep.number = Inf,
            keep.sparse = TRUE,
+           keep.these = FALSE,
            invert.selection = FALSE,
            h = NULL,
            n = NULL,
@@ -147,6 +188,7 @@ stat_dens2d_labels <-
                     keep.fraction = keep.fraction,
                     keep.number = keep.number,
                     keep.sparse = keep.sparse,
+                    keep.these = keep.these,
                     invert.selection = invert.selection,
                     h = h,
                     n = n,
@@ -161,6 +203,7 @@ dens2d_labs_compute_fun <-
            keep.fraction,
            keep.number,
            keep.sparse,
+           keep.these,
            invert.selection,
            h,
            n,
@@ -178,13 +221,32 @@ dens2d_labs_compute_fun <-
     }
 
     force(data)
-    if (nrow(data) * keep.fraction > keep.number) {
-      keep.fraction <- keep.number / nrow(data)
+    if (!exists("label", data)) {
+      message("Mapping missing 'label' aesthetic to 'rownames(data)'")
+      data[["label"]] <- rownames(data)
     }
 
-    if (!exists("label", data)) {
-      message("Mapping missing 'label' aesthetic to 'rownames(data)'.")
-      data[["label"]] <- rownames(data)
+    if (length(keep.these)) {
+      if (is.function(keep.these)) {
+        keep.these <- keep.these(data$label) # character or logical vector
+      }
+      if (is.character(keep.these)) {
+        keep.these <- keep.these == data$label # logical vector
+      }
+      if (is.numeric(keep.these)) {
+        temp <- logical(nrow(data))
+        temp[keep.these] <- TRUE
+        keep.these <- temp
+      }
+      if (anyNA(keep.these)) {
+        warning("Discarding 'NA's in keep.these")
+        keep.these <- ifelse(is.na(keep.these),
+                             FALSE,
+                             keep.these)
+      }
+    }
+    if (nrow(data) * keep.fraction > keep.number) {
+      keep.fraction <- keep.number / nrow(data)
     }
 
     if (keep.fraction == 1) {
@@ -221,11 +283,11 @@ dens2d_labs_compute_fun <-
 
     if (is.function(label.fill)) {
       if (invert.selection){
-        data[["label"]] <- ifelse(!keep,
+        data[["label"]] <- ifelse(!(keep | keep.these),
                                   data[["label"]],
                                   label.fill(data[["label"]]))
       } else {
-        data[["label"]] <- ifelse(keep,
+        data[["label"]] <- ifelse(keep | keep.these,
                                   data[["label"]],
                                   label.fill(data[["label"]]))
       }
@@ -240,11 +302,11 @@ dens2d_labs_compute_fun <-
         }
       }
       if (invert.selection){
-        data[["label"]] <- ifelse(!keep,
+        data[["label"]] <- ifelse(!(keep | keep.these),
                                   data[["label"]],
                                   label.fill)
       } else {
-        data[["label"]] <- ifelse(keep,
+        data[["label"]] <- ifelse(keep | keep.these,
                                   data[["label"]],
                                   label.fill)
       }